requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-cognitiveservices-speech==1.42.0
black==24.4.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==4.4.2
distro==1.9.0
ffmpeg-python==0.2.0
flake8==7.1.1
Flask==3.1.0
future==1.0.0
google-api-core
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-texttospeech==2.24.0
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
h11
httpcore
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
logging==0.4.9.6
MarkupSafe==3.0.2
mccabe==0.7.0
moviepy==2.1.2
mypy-extensions==1.0.0
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
openai==1.61.1
opencv-python==4.11.0.86
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
proglog==0.1.10
proto-plus==1.26.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.2.0
pyparsing==3.2.1
pytest==8.3.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
scipy==1.15.2
sentry-sdk==2.14.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.5.0
Werkzeug==3.1.3
dill==0.3.8
